Playing with Corinna: An alternate approach to a token factory
In this variation, the concrete token classes are initialized with
individual NCAs (each field a :param) instead of the token.
Mapping the arrayref data from HTML::TokeParser to key/value
constructor attributes is done in the new TokenFactory.pm.
The arrayref token is no longer used as a constructor argument.

Another, maybe preferable alternative would be to retain the token as
only :param, and leave it in the base class.  The concrete classes
would then each offer a method evaluate_token ($token) which converts
the arrayref to key/value constructor attributes, and the factory
would create objects like this:

  $class->new($class->evaluate_token($token));

evaluate_token are public methods which bloat the API in every
class, but these are class methods, completely harmless for object
privacy and integrity.